service/requirements.txt
Flask
flask-restful
pandas
numpy
Pillow
